Get data from API: Weather IPCA to Praia + Parsing

+++ app/src/main/java/mei/weatherapp/basedados/MyOpenHelper.java
@@ -122,7 +122,7 @@ public class MyOpenHelper extends SQLiteOpenHelper {
       p.setNome(cur.getString(cur.getColumnIndex(_PRAIAS_NOME_)));
       p.setLongitude(cur.getDouble(cur.getColumnIndex(_PRAIAS_LONGITUDE_)));
       p.setLatitude(cur.getDouble(cur.getColumnIndex(_PRAIAS_LATITUDE_)));
-      p.setRate(cur.getInt(cur.getColumnIndex(_PRAIAS_RATE_)));
+      p.setRate(cur.getDouble(cur.getColumnIndex(_PRAIAS_RATE_)));
       p.setTemperatura(cur.getFloat(cur.getColumnIndex(_PRAIAS_TEMPERATURA_)));
     }catch (Exception e){    }
     return p;
